Feat/test best songs

See merge request imt-conception-logicielle-leja/imt-202311-tp-spring-ezzemany!18

+++ src/test/java/imt/fil/cl/leja/songmanagerapi/singer/SingerRepositoryTest.java
@@ -1,7 +1,5 @@
 package imt.fil.cl.leja.songmanagerapi.singer;
 
-import imt.fil.cl.leja.songmanagerapi.song.Song;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
@@ -10,7 +8,8 @@ import org.springframework.test.context.jdbc.Sql;
 
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 @DataJpaTest
 @Sql("/data-singer-test.sql")
 class SingerRepositoryTest {
@@ -26,7 +25,7 @@ class SingerRepositoryTest {
     void itShouldFindSingerByIdWithBestSongs() {
         // Assume there's a singer with ID 1 and a song with rating 5 in the test data
         Long singerId = 1L;
-        int minRating = 4;
+        Float minRating = 4f;
 
         Optional<Singer> optionalSinger = singerRepository.findByIdWithBestSongs(singerId, minRating);
         // Check if the user exists
